Fixing Pug template for Last.fm API Example
Many of the URLs for images and links in the API Example for Last.fm
were incorrectly referenced, due to deprecations with attribute
interpolation syntax.

In addition, the order or promise resolution had mixed up the list
of top tracks and top albums, causing the same image to be shown for
all "top albums", and the list of "top tracks" were actually album
names. This has been corrected.