Update Readme: mLab transition to MongoDB Atlas
mLab is now part of MongoDB, Inc. and the service is being replaced in favor if MongoDB Atlas.  This commit replaces references to mLab with MongoDB Atlas, and updates the instructions.  The only exception currently is Heroku's use of mLab, since Heroku has not changed the addon and is still using their old mLab add-on through their UI.  Note that you do not need to use the Heroku Add-on and can just use env variables to point the Heroku dyno to a MongoDB Atlas instance.